refactor: Move agg-only logic out of loader
